Excutable installer and map device created
