+++ app/config/config.json
@@ -1,6 +1,6 @@
 {
     "language": "es",
-    "start_time": 1648318370,
+    "start_time": 1648374556,
     "verbose": false,
     "contexts": [
         "morning",
