+++ app/config/devices.json
@@ -1,8 +1,8 @@
 {
-    "livingLight": 1,
-    "bathroomLight": 1,
-    "bedroomLight": 1,
-    "officeLight": 1,
-    "kitchenLight": 1,
-    "hallwayLight": 1
+    "bathroomLight": 0,
+    "officeLight": 0,
+    "hallwayLight": 0,
+    "bedroomLight": 0,
+    "kitchenLight": 0,
+    "livingLight": 0
 }
