+++ app/config/temp.json
@@ -1,4 +1,4 @@
 {
-    "console_dev": "",
-    "listen": false
+    "listen": false,
+    "console_dev": ""
 }
