--- app/manual.txt
@@ -1,2 +1,0 @@
-SPACE -> open the activity menu
-Z -> Show the time formarding in the activity menu
