+++ app/openshs.py
@@ -1,36 +1,62 @@
+from tkinter import *
+from main_ui.menu_bar import *
+from main_ui.list_boxes import *
+from config.config import *
 import subprocess
-from tkinter import Tk
-from tkinter import messagebox
 import sys
 
-def main() :
+class MainUi(Tk):
+    def __init__(self):
+        super().__init__()
+        self.title("OpenSHS")
+        self.geometry("800x500")
+        self.iconbitmap("logo.ico")
+        
+        # Load custom settings for simulation
+        custom_config = Config()
+
+        self.__menu = MenuBar(self, custom_config)
+        self.config(menu=self.__menu.getMenuBar())
+
+        # Grid spaces are adjusted for components
+        self.grid_columnconfigure(0, weight=1)
+        self.grid_columnconfigure(2, weight=1)
+        self.grid_rowconfigure(1, weight=1)
+        l_boxes = ListBoxes(self, custom_config)
+
+        # MenuBar can make changes to ListBoxes
+        self.__menu.setListBoxes(l_boxes)
+
+def main():
     root = Tk()
     root.withdraw()
     try:
         with subprocess.Popen(['blender','--version'], stdout=subprocess.PIPE, universal_newlines=True) as process:
             if 'Blender 2.7' not in process.stdout.readline():
                 messagebox.showerror('Error','Blender version 2.7 must be used')
-                sys.exit(-1)
+                sys.exit()
     except: 
         messagebox.showerror('Error','No blender found on the system')
-        exit(-1)
+        exit()
 
     try:
         with subprocess.Popen(['python','--version'], stdout=subprocess.PIPE, universal_newlines=True) as process:
             if 'Python 3' not in process.stdout.readline():
                 messagebox.showerror('Error','Python 3 must be used')
-                sys.exit(-1)
+                sys.exit()
     except: 
         messagebox.showerror('Error','No python found on the system')
-        exit(-1)
-
-    
-    if subprocess.call(['pythonw','./main_ui.py'], shell=False) != 0:
-        messagebox.showerror('Error','Some program files are missing')
-        sys.exit(-1)
+        exit()
     
     try: root.destroy()
     except: pass
 
-if __name__ == '__main__' :
+    MainUi().mainloop()
+    # DEPRECATED
+    #if subprocess.call(['pythonw','./main_ui.py'], shell=False) != 0:
+    #    messagebox.showerror('Error','Some program files are missing')
+    #    sys.exit(-1)
+    
+
+if __name__ == '__main__':
     main()
