+++ app/temp/output.csv
@@ -1,42 +1,17 @@
 wardrobe,tv,oven,officeLight,officeDoorLock,officeDoor,officeCarp,office,mainDoorLock,mainDoor,livingLight,livingCarp,kitchenLight,kitchenDoorLock,kitchenDoor,kitchenCarp,hallwayLight,fridge,couch,bedroomLight,bedroomDoorLock,bedroomDoor,bedroomCarp,bedTableLamp,bed,bathroomLight,bathroomDoorLock,bathroomDoor,bathroomCarp,Activity
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,0,1,0,0,0,1,0,0,1,0,0,1,0,0,1,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,0,1,0,0,0,1,0,0,1,0,0,1,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,0,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
-0,0,1,1,1,1,1,1,0,1,0,0,0,0,1,0,1,1,1,0,0,0,1,0,0,1,0,1,0,0,0,1,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,personal
