--- help.txt
@@ -1,11 +1,0 @@
-Turn on lamp from external scripts
-logic.getCurrentScene().objects['lamp'].energy = 1.0
-logic.states['betTableLamp'] = 1
-
-Enable/disable assistant
-import json
-import os
-...
-json_object = json.dumps({'listen':True}, indent = 4)
-with open(os.path.join('config','temp.json'), "w") as json_file:
-     json_file.write(json_object)
